Move NavItem into Navigation namespace
