+++ visitz/Visitz/Views/Navigation/NavItemView.xaml
@@ -2,10 +2,10 @@
 <ContentView 
 	xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
 	xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-	xmlns:models="clr-namespace:VisitzModel.Models;assembly=VisitzModel"
+	xmlns:navModels="clr-namespace:VisitzModel.Models.Navigation;assembly=VisitzModel"
 	xmlns:nav="clr-namespace:Visitz.Views.Navigation"
 	x:Class="Visitz.Views.Navigation.NavItemView"
-	x:DataType="models:NavItem"
+	x:DataType="navModels:NavItem"
 	Padding="8">
 
 	<VisualStateManager.VisualStateGroups>
